added better callback placeholder, removed login as a class, and renamed it to dbAccess
